requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
packaging==24.0
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.1
PyYAML==6.0.1
requests==2.26.0
sqlparse==0.5.0
toml==0.10.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
